+++ src/lib.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "check-endpoints")]
+pub mod check_endpoints;
 pub mod info;
 pub mod instrumentation;
 pub mod limit_resource;